refactor(parser, parser/boolean): move explicitly boolean tests to boolean package
